+++ app/(auth)/login/page.tsx
@@ -2,8 +2,8 @@
 
 import ErrorMessage from '@/app/components/auth/ErrorMessage';
 import Image from 'next/image';
-import Input from '@/app/components/shared/input/Input';
-import Label from '@/app/components/shared/label/Label';
+import Input from '@/app/shared/input/Input';
+import Label from '@/app/shared/ui/label/Label';
 import Link from 'next/link';
 import LogoImage from '@/app/components/auth/LogoImage';
 import Title from '@/app/components/auth/Title';
